+++ Sources/NablaCore/Utils/Extensions/UIKit/UIResponder+keyboard.swift
@@ -1,0 +1,6 @@
+import Foundation
+import UIKit
+
+public extension NablaExtension where Base == UIResponder {
+    static let keyboardWillChangeFrameNotification = Notification.Name("Nabla.keyboardWillChangeFrameNotification")
+}