+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed the type of the duration in `keyboardAnimationDurationUserInfoKey` in the `keyboardWillChangeFrameNotification` notification (could cause an issue if other classes are listening and expecting a double value).   
 
 ## [1.0.0-alpha19] - 2022-10-07